[REF] account_edi_ubl_cii_*: merge modules
This commit merges `account_edi_ubl_cii_tax_extension` module into
`account_edi_ubl_cii`.

task-4926449
