+++ addons/account_edi_ubl_cii/__manifest__.py
@@ -24,6 +24,7 @@ Pro rules and show the errors.
     'depends': ['account'],
     'data': [
         'data/cii_22_templates.xml',
+        'views/account_tax_views.xml',
         'views/res_partner_views.xml',
     ],
     'assets': {
