+++ addons/account_edi_ubl_cii/models/__init__.py
@@ -10,5 +10,6 @@ from . import account_edi_xml_ubl_a_nz
 from . import account_edi_xml_ubl_sg
 from . import account_move
 from . import account_move_send
+from . import account_tax
 from . import ir_actions_report
 from . import res_partner
