--- addons/account_edi_ubl_cii_tax_extension/__init__.py
@@ -1,3 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import models
