--- addons/account_edi_ubl_cii_tax_extension/models/__init__.py
@@ -1,2 +1,0 @@
-from . import account_tax
-from . import account_edi_common
