--- addons/account_edi_ubl_cii_tax_extension/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_ubl_cii_tax_extension
