+++ addons/l10n_sg_ubl_pint/__manifest__.py
@@ -7,7 +7,7 @@
     'description': """
     The UBL PINT e-invoicing format for Singapore is based on the Peppol International (PINT) model for Billing.
     """,
-    'depends': ['account_edi_ubl_cii_tax_extension'],
+    'depends': ['account_edi_ubl_cii'],
     'installable': True,
     'author': 'Odoo S.A.',
     'license': 'LGPL-3'
